requirements.txt
django==2.2.14
djangorestframework
httpie # Realizar llamadas a la API
Faker # Generar autores con datos falsos